requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==21.4.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
clic==0.1.3
click==8.1.3
cycler==0.11.0
Cython==0.29.30
dataclasses==0.6
flatbuffers==1.12
fonttools==4.34.3
freetype-py==2.3.0
future==0.18.2
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hsluv==5.0.3
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
jsonschema==4.6.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.0
opencv-python==3.4.2.17
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.3
Pillow==8.4.0
pip==22.1.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydensecrf==1.0rc2
pyglet==1.5.26
pyglfw==0.2.2
PyOpenGL==3.1.5
pyparsing==3.0.9
pyrsistent==0.18.1
PySide6==6.2.4
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
QtPy==1.9.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.2
scipy==1.1.0
setuptools==63.1.0
shiboken6==6.2.4
six==1.15.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
tf-slim==1.1.0
tifffile==2021.11.2
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.10
vispy==0.11.0
Werkzeug==2.1.2
wheel==0.37.1
wrapt==1.12.1
zipp==3.8.0
jinja2
lmdb
ujson
msgpack